refactor: move useDate to labs
