+++ packages/vuetify/package.json
@@ -72,6 +72,7 @@
     "./locale/adapters/*": "./lib/locale/adapters/*.mjs",
     "./iconsets/*": "./lib/iconsets/*.mjs",
     "./labs/components": "./lib/labs/components.mjs",
+    "./labs/date/adapters/*": "./lib/labs/date/adapters/*.mjs",
     "./labs/*": "./lib/labs/*/index.mjs",
     "./*": "./*"
   },
