+++ packages/vuetify/src/composables/index.ts
@@ -3,14 +3,12 @@
  * Imports in our code should be to the composable directly, not this file
  */
 
-export { useDate } from './date'
 export { useDefaults } from './defaults'
 export { useDisplay } from './display'
 export { useLayout } from './layout'
 export { useLocale, useRtl } from './locale'
 export { useTheme } from './theme'
 
-export type { DateOptions, DateInstance } from './date'
 export type { DefaultsInstance } from './defaults'
 export type { DisplayBreakpoint, DisplayInstance, DisplayThresholds } from './display'
 export type { SubmitEventPromise } from './form'
