+++ packages/vuetify/src/labs/date/__tests__/date.spec.ts
@@ -7,7 +7,7 @@ import { mount } from '@vue/test-utils'
 import { describe, expect, it } from '@jest/globals'
 
 // Types
-import VuetifyDateAdapter from '@/adapters/vuetify'
+import { VuetifyDateAdapter } from '../adapters/vuetify'
 import { defineComponent, h } from 'vue'
 
 describe('date.ts', () => {
