+++ packages/vuetify/src/labs/date/index.ts
@@ -1,0 +1,4 @@
+export { useDate } from './date'
+
+export type { DateAdapter } from './DateAdapter'
+export type { DateOptions, DateInstance } from './date'
